// front-end/src/client/models/StudentUpdateParams.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type StudentUpdateParams = {
    name?: (string | null);
    sex?: ('M' | 'F' | null);
    age?: (number | null);
    current_campus?: ('A' | 'B' | 'C' | null);
};